mobile_api/requirements.txt
black==22.1.0
click==8.0.3
Django==2.2.19
django-filter==21.1
djangorestframework==3.13.1
flake8==4.0.1
gunicorn==20.1.0
Markdown==3.3.6
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2021.3
sqlparse==0.4.2
tomli==2.0.1

